Create crate::cli::options
