+++ src/cli/mod.rs
@@ -1,4 +1,5 @@
 pub mod directory_check;
+pub mod options;
 pub mod self_install;
 pub mod self_migrate;
 pub mod self_upgrade;
