+++ src/commands/cli.rs
@@ -4,7 +4,8 @@ use crate::cli::directory_check;
 use crate::cli::self_install;
 use crate::cli::self_migrate;
 use crate::cli::self_upgrade;
-use crate::options::{Options, Command, SelfSubcommand};
+use crate::options::{Options, Command};
+use crate::cli::options::SelfSubcommand;
 use crate::commands::parser::{Common, MigrationCmd, Migration};
 use crate::non_interactive;
 use crate::commands;
